// src/app/data/model/Edamam.ts
export interface Ingredient {
    text: string;
    weight: number;
}

export interface ENERC_KCAL {
    label: string;
    quantity: number;
    unit: string;
}

export interface FAT {
    label: string;
    quantity: number;
    unit: string;
}

export interface FASAT {
    label: string;
    quantity: number;
    unit: string;
}

export interface FATRN {
    label: string;
    quantity: number;
    unit: string;
}

export interface FAM {
    label: string;
    quantity: number;
    unit: string;
}

export interface FAPU {
    label: string;
    quantity: number;
    unit: string;
}

export interface CHOCDF {
    label: string;
    quantity: number;
    unit: string;
}

export interface FIBTG {
    label: string;
    quantity: number;
    unit: string;
}

export interface SUGAR {
    label: string;
    quantity: number;
    unit: string;
}

export interface PROCNT {
    label: string;
    quantity: number;
    unit: string;
}

export interface CHOLE {
    label: string;
    quantity: number;
    unit: string;
}

export interface NA {
    label: string;
    quantity: number;
    unit: string;
}

export interface CA {
    label: string;
    quantity: number;
    unit: string;
}

export interface MG {
    label: string;
    quantity: number;
    unit: string;
}

export interface K {
    label: string;
    quantity: number;
    unit: string;
}

export interface FE {
    label: string;
    quantity: number;
    unit: string;
}

export interface ZN {
    label: string;
    quantity: number;
    unit: string;
}

export interface P {
    label: string;
    quantity: number;
    unit: string;
}

export interface VITA_RAE {
    label: string;
    quantity: number;
    unit: string;
}

export interface VITC {
    label: string;
    quantity: number;
    unit: string;
}

export interface THIA {
    label: string;
    quantity: number;
    unit: string;
}

export interface RIBF {
    label: string;
    quantity: number;
    unit: string;
}

export interface NIA {
    label: string;
    quantity: number;
    unit: string;
}

export interface VITB6A {
    label: string;
    quantity: number;
    unit: string;
}

export interface FOLDFE {
    label: string;
    quantity: number;
    unit: string;
}

export interface FOLFD {
    label: string;
    quantity: number;
    unit: string;
}

export interface FOLAC {
    label: string;
    quantity: number;
    unit: string;
}

export interface VITB12 {
    label: string;
    quantity: number;
    unit: string;
}

export interface VITD {
    label: string;
    quantity: number;
    unit: string;
}

export interface TOCPHA {
    label: string;
    quantity: number;
    unit: string;
}

export interface VITK1 {
    label: string;
    quantity: number;
    unit: string;
}

export interface WATER {
    label: string;
    quantity: number;
    unit: string;
}

export interface TotalNutrient {
    eNERC_KCAL: ENERC_KCAL;
    fAT: FAT;
    fASAT: FASAT;
    fATRN: FATRN;
    fAMS: FAM;
    fAPU: FAPU;
    cHOCDF: CHOCDF;
    fIBTG: FIBTG;
    sUGAR: SUGAR;
    pROCNT: PROCNT;
    cHOLE: CHOLE;
    nA: NA;
    cA: CA;
    mG: MG;
    k: K;
    fE: FE;
    zN: ZN;
    p: P;
    vITA_RAE: VITA_RAE;
    vITC: VITC;
    tHIA: THIA;
    rIBF: RIBF;
    nIA: NIA;
    vITB6A: VITB6A;
    FOLDfE: FOLDFE;
    fOLFD: FOLFD;
    fOLAC: FOLAC;
    vITB12: VITB12;
    vITD: VITD;
    tOCPHA: TOCPHA;
    vITK1: VITK1;
    wATER: WATER;
}
// ...
export interface TotalDaily {
    eNERC_KCAL: ENERC_KCAL;
    fAT: FAT;
    fASAT: FASAT;
    cHOCDF: CHOCDF;
    fIBTG: FIBTG;
    pROCNT: PROCNT;
    cHOLE: CHOLE;
    nA: NA;
    cA: CA;
    mG: MG;
    k: K;
    fE: FE;
    zN: ZN;
    p: P;
    vITA_RAE: VITA_RAE;
    vITC: VITC;
    tHIA: THIA;
    rIBF: RIBF;
    nIA: NIA;
    vITB6A: VITB6A;
    FOLDfE: FOLDFE;
    vITB12: VITB12;
    vITD: VITD;
    tOCPHA: TOCPHA;
    vITK1: VITK1;
}

export interface Sub {
    label: string;
    tag: string;
    schemaOrgTag: string;
    total: number;
    hasRDI: boolean;
    daily: number;
    unit: string;
}

export interface Digest {
    label: string;
    tag: string;
    schemaOrgTag: string;
    total: number;
    hasRDI: boolean;
    daily: number;
    unit: string;
    sub: Sub[];
}

export interface Recipe {
    uri: string;
    label: string;
    image: string;
    source: string;
    url: string;
    shareAs: string;
    yield: number;
    dietLabels: string[];
    healthLabels: string[];
    cautions: string[];
    ingredientLines: string[];
    ingredients: Ingredient[];
    calories: number;
    totalWeight: number;
    totalTime: number;
    totalNutrients: TotalNutrient;
    totalDaily: TotalDaily;
    digest: Digest[];
}

export interface Hit {
    recipe: Recipe;
    bookmarked: boolean;
    bought: boolean;
}